REFACTOR :: PostAppendButton 파일분할
